move console stuff from loader
